+++ mitra_models/src/properties/constants.rs
@@ -1,2 +1,3 @@
+pub const FEDERATED_TIMELINE_RESTRICTED: &str = "federated_timeline_restricted";
 pub const ETHEREUM_CURRENT_BLOCK: &str = "ethereum_current_block";
 pub const ETHEREUM_TOKEN_WAITLIST_MAP: &str = "token_waitlist_map";